# docs/api/openapi-bundled.yaml
openapi: 3.1.0
info:
  title: AuthFramework REST API
  description: |
    Comprehensive REST API for the AuthFramework authentication and authorization library.

    This API provides complete access to AuthFramework's authentication, authorization,
    user management, multi-factor authentication, OAuth 2.0, OIDC, RBAC, and administrative 
    features through standardized HTTP endpoints.

    ## Version Strategy

    - **Versioned endpoints** (`/api/v1/*`): All functional API endpoints are versioned
      for stability and backward compatibility. Version changes follow semantic versioning.

    - **Unversioned endpoints**: Infrastructure endpoints (`/health`, `/metrics`, `/readiness`, 
      `/liveness`) and OIDC discovery (`/.well-known/*`) remain unversioned for operational 
      stability and spec compliance.

    ## Authentication

    Most endpoints require authentication via Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```

    Obtain tokens via:
    - `/api/v1/auth/login` - Username/password authentication
    - `/api/v1/oauth/token` - OAuth 2.0 token endpoint
    - `/api/v1/auth/refresh` - Refresh token flow

    ## Rate Limiting

    API endpoints are rate-limited to prevent abuse:
    - Authentication endpoints: 5 requests per minute per IP
    - Standard endpoints: 100 requests per minute per user
    - Admin endpoints: 50 requests per minute per admin user

...

    ## Response Format

...
    ```json
    {
...
    }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/oauth/authorize:
    get:
      tags:
        - OAuth
      summary: OAuth 2.0 authorization endpoint
      description: |
        OAuth 2.0 authorization endpoint (RFC 6749). Initiates authorization code flow 
        or implicit flow. Supports PKCE (RFC 7636) for public clients.
      operationId: authorize
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              - code
              - token
              - id_token
          description: OAuth response type
          example: code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: OAuth client identifier
          example: client_123
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Redirect URI (must be pre-registered)
          example: https://client.example.com/callback
        - name: scope
          in: query
          schema:
            type: string
          description: Requested scope (space-separated)
          example: openid profile email
        - name: state
          in: query
          schema:
            type: string
          description: Opaque state value for CSRF protection
          example: xyz123
        - name: code_challenge
          in: query
          schema:
            type: string
          description: PKCE code challenge
          example: E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum:
              - S256
              - plain
          description: PKCE code challenge method
          example: S256
        - name: nonce
          in: query
          schema:
            type: string
          description: OIDC nonce for replay protection
          example: n-0S6_WzA2Mj
      responses:
        '302':
          description: Redirect to client redirect_uri with authorization code or error
          headers:
            Location:
              schema:
                type: string
              description: Redirect URI with code/error
              examples:
                success:
                  value: https://client.example.com/callback?code=abc123&state=xyz123
                error:
                  value: https://client.example.com/callback?error=invalid_request&error_description=Missing+client_id&state=xyz123
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/oauth/userinfo:
    get:
      tags:
        - OAuth
        - OIDC
      summary: OpenID Connect UserInfo
      description: |
        OIDC UserInfo endpoint (OIDC Core 5.3). Returns claims about the authenticated user.
        Requires valid access token with 'openid' scope.
      operationId: userinfo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Invalid or missing access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: invalid_token
                  error_description:
                    type: string
                    example: The access token is invalid or expired
        '403':
          description: Insufficient scope (openid scope required)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: insufficient_scope
                  error_description:
                    type: string
                    example: The request requires 'openid' scope
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-access-token>
        ```
...
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            validationError:
              summary: Validation error example
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: Invalid input parameters
                  details:
                    field: email
                    issue: Invalid email format
                timestamp: '2024-01-20T15:30:45Z'
...
  schemas:
...
    ApiError:
      type: object
      description: Standard error response wrapper
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: object
          description: Error details
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: INVALID_CREDENTIALS
            message:
              type: string
              description: Human-readable error message
              example: The provided credentials are invalid
            details:
              type: object
              nullable: true
              description: Additional error context (structure varies by error type)
          required:
            - code
            - message
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the error
          example: '2024-01-20T15:30:45Z'
      required:
        - success
        - error
        - timestamp
...
    UserInfoResponse:
      type: object
      description: OpenID Connect UserInfo response (OIDC Core 5.3)
      properties:
        sub:
          type: string
          description: Subject identifier (user ID)
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          description: Full name
          example: John Doe
        given_name:
          type: string
          description: First name
          example: John
        family_name:
          type: string
          description: Last name
          example: Doe
        email:
          type: string
          format: email
          description: Email address
          example: user@example.com
        email_verified:
          type: boolean
          description: Whether email has been verified
          example: true
        phone_number:
          type: string
          description: Phone number
          example: '+1234567890'
        phone_number_verified:
          type: boolean
          description: Whether phone number has been verified
          example: false
        picture:
          type: string
          format: uri
          description: Profile picture URL
          example: https://example.com/avatars/user123.jpg
        locale:
          type: string
          description: User's locale
          example: en-US
        zoneinfo:
          type: string
          description: User's timezone
          example: America/New_York
        updated_at:
          type: integer
          description: Last profile update timestamp (Unix epoch)
          example: 1642686645
      required:
        - sub
...